+++ src/main/java/calculator/parser/MathematicalExpressionParser.java
@@ -2,5 +2,5 @@ package calculator.parser;
 
 public interface MathematicalExpressionParser {
 
-	public int[] parse(String text, String separatorRegex);
+	public long[] parse(String text, String separatorRegex);
 }
